requirements.txt
pandas>=2.1.4
numpy>=1.26.4
scikit-learn>=1.4.0
xgboost>=2.0.3
yfinance>=0.2.36
plotly>=5.18.0
joblib>=1.3.2
python-dotenv>=1.0.1
streamlit>=1.31.1
requests>=2.31.0
